+++ src/Storages/DistributedMergeTree/StreamingBlockReader.h
@@ -20,7 +20,7 @@ public:
         std::shared_ptr<IStorage> storage_,
         Int32 shard_,
         Int64 offset,
-        std::vector<size_t> column_positions,
+        std::vector<uint16_t> column_positions,
         const DWAL::KafkaWALSimpleConsumerPtr & consumer_,
         Poco::Logger * log_);
